chore: move types and typings to /src
