+++ .github/workflows/deployment.yml
@@ -1,4 +1,4 @@
-name: "🚀 Deployment"
+name: Deployment
 
 on:
   push:
