+++ src/components/dashboard/sidebar/Sidebar.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { memo } from "react";
 
-import type { HeroIcon } from "../../../../types/common.types";
+import type { HeroIcon } from "../../../types/common.types";
 
 interface NavigationItem {
   readonly name: string;
