+++ src/views/login/Login.tsx
@@ -8,13 +8,13 @@ import { memo, useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import toast from "react-hot-toast";
 
-import { AUTH_PROVIDER } from "../../../types/auth.types";
 import { Button } from "../../components/common/Button";
 import { Input } from "../../components/common/Input";
+import { AUTH_PROVIDER } from "../../types/auth.types";
 import { onPromise } from "../../utils/functions";
 import { loginUserSchema } from "../../utils/validation/schema";
 
-import type { Login } from "../../../types/auth.types";
+import type { Login } from "../../types/auth.types";
 import type { ClientSafeProvider } from "next-auth/react";
 
 interface LoginProps {
